3교시중간커밋

+++ src/main/java/org/edu/controller/AdminController.java
@@ -89,26 +89,45 @@ public class AdminController {
 		return "admin/board/board_update";//파일경로
 	}
 	@RequestMapping(value="/admin/board/board_update",method=RequestMethod.POST)
-	public String board_update(RedirectAttributes rdat,MultipartFile file, BoardVO boardVO, PageVO pageVO) throws Exception {
+	public String board_update(RedirectAttributes rdat,@RequestParam("file") MultipartFile[] files, BoardVO boardVO, PageVO pageVO) throws Exception {
 		//기존 등록된 첨부파일 목록 구하기
 		List<HashMap<String,Object>> delFiles = boardService.readAttach(boardVO.getBno());
+		String[] save_file_names = new String[files.length];
+		String[] real_file_names = new String[files.length];
+		int index = 0;//아래 향상된 for문에서 사용할 인덱스값
 		//첨부파일 수정: 기존첨부파일 삭제 후 신규파일 업로드
-		if(file.getOriginalFilename() != "") {//첨부파일명이 있으면
-			//기존파일 폴더에서 삭제 처리
-			for(HashMap<String,Object> file_name:delFiles) {
-				File target = new File(commonController.getUploadPath(), (String) file_name.get("save_file_name"));
-				if(target.exists()) {
-					target.delete();//폴더에서 기존첨부파일 지우기
-					//서비스클래스에는 첨부파일DB를 지우는 메서드가 없음. DAO를 접근해서 tbl_attach를 지웁니다.
-					boardDAO.deleteAttach((String) file_name.get("save_file_name"));
+		for(MultipartFile file:files) {//다중파일 업로드 호출 부분 시작 향상된 for문 사용
+			if(file.getOriginalFilename() != "") {//첨부파일명이 있으면
+				//기존파일DB에서 삭제처리할 변수 생성
+				int cnt = 0;
+				for(HashMap<String,Object> file_name:delFiles) {
+					save_file_names[cnt] = (String) file_name.get("save_file_name");
+					real_file_names[cnt] = (String) file_name.get("real_file_name");
+					cnt = cnt + 1;//반복시 증가값	
+					}
+				int sun = 0;//업데이트jsp에서 첨부파일을 개별삭제할때시 사용할 
+				//기존파일 폴더에서 삭제 처리
+				for(HashMap<String,Object> file_name:delFiles) {
+					if(index == sun) {//index는 첨부파일개수, sun삭제할 개별순서
+						
+					
+						File target = new File(commonController.getUploadPath(), (String) file_name.get("save_file_name"));
+						if(target.exists()) {
+							target.delete();//폴더에서 기존첨부파일 지우기
+							//서비스클래스에는 첨부파일DB를 지우는 메서드가 없음. DAO를 접근해서 tbl_attach를 지웁니다.
+							boardDAO.deleteAttach((String) file_name.get("save_file_name"));
+							sun = sun + 1;//개별삭제는 for문에서 딱1번뿐이기 때문에
+						}
+					}
 				}
+				//신규파일 폴더에 업로드 처리
+				save_file_names[index] = commonController.fileUpload(file);//폴더에 업로드저장완료				
+				real_file_names[index] = file.getOriginalFilename();//"한글파일명.jpg"				
 			}
-			//신규파일 폴더에 업로드 처리
-			String[] save_file_names = commonController.fileUpload(file);//폴더에 업로드저장완료
-			boardVO.setSave_file_names(save_file_names);//UUID로 생성된 유니크한 파일명
-			String[] real_file_names = new String[] {file.getOriginalFilename()};//"한글파일명.jpg"
-			boardVO.setReal_file_names(real_file_names);
+				index = index + 1;
 		}
+		boardVO.setSave_file_names(save_file_names);//UUID로 생성된 유니크한 파일명
+		boardVO.setReal_file_names(real_file_names);
 		boardService.updateBoard(boardVO);//DB에서 업데이트
 		rdat.addFlashAttribute("msg", "수정");
 		return "redirect:/admin/board/board_view?page="+pageVO.getPage()+"&bno="+boardVO.getBno();
